chore(ci): allow codegen to proceed if there is an error during patching
Allowing codegen to proceed in the face of failed spec patching makes it
easier to pull in an upstream spec that includes local patches.  This has
the trade-off that failed patches _may_ lead to unwanted/unexpected changes
to the generated code, but this can be caught in review.  The sync workflow
still puts the PR in draft mode if patching fails to act as a speedbump when
sync workflow results need more attentive review.